+++ src/ios/functions.js
@@ -1,4 +1,3 @@
-const inquirer = require("inquirer");
 const { exec } = require("child_process");
 const { exit } = require("process");
 const {
